refactor(integration): adobe analytics - remove productMerchEventToAdobeEvent dependency, refactor to util
